Added a print preview button, and have the basics in place for selecting different templates.  It still needs work, but I know what I need to do.
I rearranged the brewday instructions screen a lot to make room for new
buttons.  I am not sure if this is a good thing or not.

I also added a calorie calculator.  Right now, it is only shown on the print
pages, simply because I couldn't find a good place on the main recipe display
to put it.