I see your tabbed widget and raise you one print and print preview menu.
The recipe formatter is done.  The copy to clipboard still works as it did.  I
have simply used that as a template for putting together the CSS+HTML
view.  The print and printpreview functions work pretty much like they do for
the brew day.

I have removed the print buttons from the brewDayWidget and the
brewDayScrollWidget.  I have also momentarily removed the CSS selection
mechanism.  I want that to be another menu option, but don't have time to
figure it all right now.

I have also updated the brewday print mode to include any existing notes in
the print out.

This should close 3043803.  Please note, it does not address 3156743