Initial checkin for brewnotes.
The base functionality is provided by two new classes: BrewNoteWidget and
BrewNote. The brewnotes are part of the recipe as I have implemented them,
which accounts for the changes to recipe.cpp.  I also added a method
getPoints() to aid with some of the work in brewnotes.  

The Tree item, model and view needed slight modifications to properly display
and manipulate the notes.  Minor improvements have been made to the
BrewTargetTreeModel::notifier to update the tree view when a recipe is
modified.

MainWindow underwent a large amount of changes to support the new layer.  The
context menus for the brewnotes have been defined, and a new action on a
recipe for brewing it is defined.