Fixes bug 3304910
It was a more fundamental problem than I had realized.  It caused me to
revisit some of the initial design elements.

The brewnote object now remembers who its parent recipe is, which changes
BrewTargetTreeModel significantly, for the better.. It uses a QDateTime for
the brewDate now in order to keep multiple brews on one day straight.  The
QDateTime is used almost everywhere to percisely identify one brew day.  I
added getBrewDate_short to get the yyyy-mm-dd format used for display
purposes.