Closes feature request 3297712.
Deleting things will now pop a question window to verify the user wants to
delete.  "Yes", "Yes to all", "No" and "cancel" are the available options.  if
the user says "Yes to all", no further prompts are made.

This checkin also tries to make more sense of how the tree changes whats
selected.  Deleting anything will end up with the first recipe being selected
and displayed.  Opening a new brewnote will cause that note to be selected and
the parent tree to be expanded.