Closes bug 3370570.
In one of my refactorings, I added the setSelection() method.  You need to
call setRecipe() (or an equivalent) for this to work.  I did it everywhere
but newRecipe().  Now I do it in newRecipe() as well.  I am still not quite
certain why the bug manifested itself the way it did, with only the mash
getting wiped.  I will guess pointer magic.