Trying to squash bug 3370570.
It turns out that on some distros, the mashComboBox is getting an
currentIndexChanged signal when a recipe is copied or added. This results in
the mash of the recipeObs to be changed.  The bad part is that this is
happening before the recipeObs it pointed to the new recipe.  

The root cause seems to be that the Database::addRecipe() method issues a
hasChanged(DBALL) at the bottom, which is causing the mashComboBox to reload
itself and this seems to be issuing said signal.

I have modified the addRecipe() method to use hasChanged(DBRECIPE) instead. It
fixes the problem, it doesn't appear to break anything else and even makes
creating a new recipe blazing fast.

I just have no idea what this will break nor why this problem happened on
Slackware and Debian Wheezy, but not gentoo or Debian Squeeze.