Closing feature request 3315501.
Brewnotes are now included with the recipe print outs. This includes the
preview, the print and the export to html functions. It isn't pretty, but it
isn't any uglier than what I am doing now. One of these days, it would be nice
to get somebody who can actually speak css and html to rewrite these for me.

I cleaned up the html a little bit as well. I found some places where the tr
tags weren't closed, and the div tag wasn't closed.

Finally, I fixed a small bug in the brewnote that resulted in the final volume
not being preserved.