Closes bug 3404030
It was an array index out of bounds issue caused by using an unsigned int
equal to 0, subtracting one and then checking to see if the result is less
than 0. Notice the unsigned part?

I found a similar bug in the the BrewDay Scroll widget too, so I fixed that as
well.