More work on ticket 3404019.
Rev 550 was far too simplistic. It's main problem was that every change was
made to the object in memory. If the user pressed the "CANCEL" button, it
should revert to whatever the user did while the window was open. This change
brings us closer to that idea, but it has some flaws; specifically, it doesn't
play well with the dirtybit and there isn't much I can do about that right
now.