Complete refactoring of the tree windows.
This is a pretty big change to the layout. I have replaced the one tree with
multiple trees in a tabbed container. I believe this should work just like the
old tree did, but now it will give us the chance to sort. The implementation
is pretty fancy, with inheritance, polymorphism and the like.