More schema ideas
The last idea of using the name and version wasn't well received, so this is
the next idea. The child/parent relationship is now in a separate table. I am
not sure of this idea either because finding the descendants of an ingredient is a
little cumbersome, but that shouldn't hopefully be too common.

You will need to drop any existing database and reload. The instructions are
still the same:
   sqlite3 -init create.sql brew.db
   perl parse_database.pl -p ~/.config/brewtarget