Keyboard can now be used to select items in the tree view.
Fixes issue 3435864

The original implementation ignored the keys because the only signal I could
find was 'activate', which fires when the item is clicked as well. This is bad
interface design, because you should be able to do everything from the
keyboard that you can from the mouse. I fixed it by implementing the
keyPressEvent handler. It simply looks to see if the key is a space or enter
from either the keyboard or keypad. If it is, it sends a doubleClick signal.