Better sorting for hops and fermentables in recipe outputs.
Closes feature request 3453098

In the recipe formatter output, this patch will list the fermentables 
in descending order by weight. The hops will be sorted by use (mash, first
wort, boil, aroma and dry hops) and then by descending order of time.

It adds:
   two public methods: Recipe::sortHopsByTime() and Recipe::sortFermentablesByWeight()
   two private methods: Recipe::hopLessThanByTime() and Recipe::fermentableLessThanByWeight()
It also promotes the private method Recipe::ibuFromHop() to a public method. 

RecipeFormatter::buildFermentableTable() changes slightly to use the sorted
list generated by sortFermentablesByWeight() instead of the fermentables().

Same for RecipeFormatter::buildHopsTable(). ibuFromHop() had to be made public
for buildHopsTable() to properly display the ibus of each hop. Prior to this
patch, buildHopsTable() relied on the fact the recipe->hops() and
recipe->IBUs() were in the same order. Once we sorted the hops list, that
relationship was broken and we have to figure the ibu for each hop.

Finally, HopSortFilterProxyModel::lessThan was updated to use the sort
algorithm as shown above when the user selects the "time" column. It really
makes more sense.