Sorry for the multiple changes on one checkin, but they are all intertwined.
The schema had to be updated to allow the instructions table to load. It would
seem that autoincrement only works in sqlite for primary keys. I removed the
autoincrement and added a trigger to get the same job done. This meant I had
to change the parsing in database.cpp to break on \n\n instead of ; and that
meant I had to remove the excess blank lines.

Updated instruction.cpp to make the properties and database fields match.

Updated the properties in recipe.cpp to match the changes made in the schema.

Added some logic to BeerXMLElement::getDate to attempt to handle the various
date formats we have used.

Added the flush() call to recipeFromXml. Modified the call to
equipmentFromXml() in recipeFromXml() to use the entire node, not the first
element. brewnotesFromXml() is currently disabled until I get other things
fixed.