Database::copy() was attempting to update the id for a row in a table to a previously existing value. This was throwing an error that wasn't reported until the foreign_keys pragma was enabled. Once this was fixed, the code appears to work and has stopped inserting empty rows.
Once this problem was resolved, it became apparent that
Database::addIngredientToRecipe() wasn't using the right key name when
inserting records into the linking tables.

The pragmas have been changed to enable journaling and to enable asynchronous
writes to the database. Disabling journaling also disabled the transactions,
which is not necessarily good. Enabling the asynchronous writes significantly
improved the performance to offset the penalty for enabling the journals. It
should be noted that doing this poses the risk of the OS (or hard drive, or
...) crashing before the filesystem can flush the writes to disk. If this
happens, the database will be corrupted.

The flush() command had to be pushed into each of the FromXml() methods in
order for those ingredients to be available for addToRecipe(). 

Finally, there are some prototypes for finer control of importing ingredients
from XML. Basically, any hop, fermentable, etc. read while importing recipes
is already a copy by definition. Instead of cluttering the DB with a bunch of
duplicated items, the code will know when it needs to copy and when it
doesn't. Right now, this is only implemented for hops but the next checkin
will fully implement it.