The trees will now obey the display and deleted flags.
Closes bug ID 3508315

But it likely opens more. The trees will now only show the items marked as
display=true and delete=false. We likely need to fix that so that the user can
select what they want to see -- normal, normal + all deleted or normal + all.
I am a little uncomfortable with my changes; it should have been done as a
sort filter/proxy instead of abusing the model directly. I reserve the right
to implement this correctly just as soon as I find it is dumping core.

I had to fix BeerXMLElement to set the fields to 0 or 1, not TRUE or FALSE.
This necessitated some changes in the create_db.sql file as well. Finally, the
database.h had to be modified to properly set the display() attribute.