Added fetchMore() to populateElements() to make sure we grab everything.
Closes bug ID 3509082

It would seem that QSqlQueryModel::select() will only return the first 256
rows of a query when the database doesn't tell it how many rows will be
returned. sqlite is a database that doesn't tell how many rows a query will
return.

To fix this, I added a tight while-loop into database::populateElements that
calls fetchMore() as long as there is more to fetch. This seems to fix the
problem.