Removes the signal warning on start up, makes sure deleting all recipes doesn't sigsegv
Closes bug ID 3511121

Sorry for the multiple bugs on one checkin, but they got mixed together and I
didn't want to untangle them. The warning on startup for the MashEditor signal
is gone. I also make sure that if you delete all the recipes, we don't attempt
to set focus on a non-existing recipe.