Proof of concept pass at allowing users to select which unit (e.g., metric, US or British Imperial) will be used to display a specific field. This is POC only and only works for one field right now.
It needs to be made better. The menus shouldn't be generated until required,
if only to reduce the amount of work done at start up. The menu needs to
indicate what the current selection is. I am not 100% sold on how I'm doing
this. Should we use the dynamic properties on teh field? Its elegant, but
hard to extend w/o a lot of manual code. A hash?