Next pass at unit magic. Requires a drop of the database and rebuild.
First, you can select the unit displayed in the fermentables table. This works
like the previous checkin -- right click an amount field and select your
preference. What's fun is that it is preserved over restarts, and only ever
effects that row. Don't care anymore? Select "default" and it will revert to
the system default. Anything that hasn't been explicitly selected follows the
default.

Second, you can select the scale. Don't want to see 8 oz? Fine. Select a scale
of pounds and you will see 0.5lbs. Want to see milligrams instead of
kilograms? That works too. Want to revert to the default? Select that. 

The deep magic only works for the fermentables table right now. The selecting
"default" needs to be backported to the BtLabels.