Selecting scale and units continues to evolve.
I have wired up -- still on the fermentables table only -- the column header.
If you right-click, you will get a menu allow you to select the units and
scale for the entire column.  The cells still work too. If you select a unit
or scale, it will over ride the column's settings. If you change the column
selection, it will revert all cells to default settings.

This is still a prototype. As discussed in the email chain, there are still
outstanding elements I need to examine and fix -- notably the calls to
displayAmount(). They look prettier in the tables because I can make more
assumptions.