Unit and scale makes progress.
I think I have the underlying methods correct. The calls aren't too bad and
seem to be working. 

Most of the brewnote display works, but there are a few odd bugs on the
postboil page I need to go chasing.  I still need to get temperature working
on the brewnotes.

I need to get color working everywhere. I need to get all of the tables wired
up.  I need to look at my implementation a bit more. There is getting to be a lot
of cut'n'paste code between MainWindow and BtLabel, which tells me I have a
design issue.