Made additional updates to the yeast pitching calculator.  Added aeration, viability, and pitching counts to the equation.
In terms of fixes, the slider SIGNAL was changed to valueChanged.  This fixes a problem where dragging the slider back and forth was resulting in inconsistent calculations.  Also, the previous SIGNAL did not fire if the user changed the value by clicking on the slider.  The user had to always drag it in order to update calculations.