Small performance improvement.
YeastDialog::populateTable() and MiscDialog::populateTable() were both
iterating through the list of yeast and misc (respectively) and adding them to
the table model one at a time. This generated a lot of work, since inserting
one at a time caused the beginInsertRow() and endInsertRow() to be called for
each item, and then the table rows (which are not visible at this point!) to
be resized.

The populateTable() methods were modified to use the bulk load methods
addYeasts() and addMiscs() (respectively), which call the beginInsertRow and
endInsertRow only once and resize the table columns only after all the data
has been loaded.

The startup time has gone from 15 seconds to 2 seconds.