Closes bug id 3547337 and 3547336
There were a slew of issues with the brewnotes. This checkin fixes as many of
them as I could find.

For 3547336, I have added a key() method to the BrewNote so I can access the
underlying database key. I needed something unique, and that was the easiest
thing I could find. I changed the brewNotes MultiHash to a Hash in MainWindow,
because now I have guaranteed uniqueness in the keys. This made a lot of bugs
go away. I don't think deleting brewnotes in bulk ever really worked. Now it
should. I also discovered the brewnote queries weren't exactly playing nice
with the delete flag. I cleaned that up everywhere it made sense.

For 3547336, the problem was far more subtle. It was essentially that the
MainWindow was adding a brewnote directly to the database, which caused the
database to generate the signal and there was nothing in the tree model to
catch that signal. The problem was compounded by the fact that no matter where
the brewnote was created, it would always be the database that generated the
signal and the tree model had no idea what recipe the brewnote was associated
with. I fixed this by resurrecting the addBrewNote() method in recipe, which
is what handles creating and populating the brewnote. I modified the database
methods to accept a parameter to determine if the database generated the
signal or not. Adding a brewnote now has a signal generated by the recipe,
which makes it all work.

I found a few other minor issues with the signalling that I cleaned up.