Closes bug 3558604.
In-memory scratch pads don't work anymore, and getting the reagents list
for heating the water and adding the grains required it. I created three new
methods on the recipe to just generate the list of reagents. These methods are
used to generate the instruction's directions, but allows BrewDayScrollWidget
to access the lists too.

I also fixed a small bug in initializing the BrewDayScrollWidget that
was preventing the print preview window from showing.