Closes bug id 3564884.
Nothing was causing the calculated fields to be recalculated when an input
changed. For example, changing the measured preboil gravity wasn't causing
the efficiency into boil kettle to be recalculated.

This maybe should have been done with signals/slots, but only the Brewnote
that generated the signal would have cared about. I have simply updated the
affected setters to cause the recalculations.