Closes bug ID 3545035
The automagic guessing of what units you entered now works. For example, if
the default unit for volume is US, but you have set the volume field in the
mashsteps to SI, entering just "20" is interpreted as 20 liters, not 20
gallons. To the best of my ability to determine, this works for any field that
it makes sense for it to work for.

This pretty much closes out the unit and scale work.