Closes bug ID 3563844
We started treating styles like other ingredients and making copies. This
causes a variety of different issues, most of which this checkin cleans up.
The UI has been reworked slightly to make the style work like the equipment.
The SQL interfaces have been modified to behave correctly. The style editor
has been modified to behave like the equipment editor. 

In addition, a problem with unquoted ' was dealt with. Maybe not well, but
changing the style should work.