Tree views were not refreshing properly when items were deleted. It had to do with how we are refactoring the signals.
Since I now know what has changed, I can be a lot more elegant in how it is
handled. I only remove the rows that need removing based on what was
deleted, or I just add a new row depending on what was inserted. This makes it
so much more clean. The outstanding item is if somebody renames an ingredient.
I will need to see how that works.