Closes 3582083
This closes really two bugs. I have been able to reduce the load time on my
database from 1m40s to about 18s by modifying when the signals are generated
and how those signals are handled. 

Previous to this check in, signals were being generated after an ingredient
was created, but before the data was populated. This caused two issues. First,
the trees redrew themselves for each element being imported. Second, this
caused problems with the trees showing elements they shouldn't be showing. The
problem was fixed by blocking the signals until after the data was populated
and causing the trees to do nothing when the ingredient wasn't to be
displayed.

I also fixed a few minor annoyances with setRgb.