Closes bug 3588867
Recipe::points was using the og() to calculate the total number of points.
This number already took into account the expected efficiency. The brewnote
calculations assumed this was the theoretical maximum available, not the
already adjusted amount. It was also impossible to go from the OG back to the
total theoretical amount.

I split the sugars calculations out into their own method in Recipe, and then
modified the calculations in brewnote to finish the math. Everything seems to
be working better.