Closes bug 3588868
Modified the MainWindow to use a more complex algorithm for determining
whether to import the defaults or not.

This change solves four potential cases: 
   a user brand new to brewtarget, using v2.0
   a brewtarget v1.2.5 user using v2.0 for the first time
   the developers, with crikey only knows what kind of cruft
   an already converted user

This done by adding a "converted" flag to the configuration file. If that flag
is set, nothing is done and brewtarget starts as expected. This solves the
fourth case.

If the flag isn't set, check for the existence of the obsolete directory. If
it exists, we interpret this to mean this is a developer and nothing more
needs to be done. The converted flag is set and brewtarget starts as normal.
This resolves the third use case.

If the obsolete directory doesn't exist, we check for the existence of the old
XML files. If they exist, we import those XML files and then set the converted
flag. This solves the second use case.

Finally, if the old XML files do not exist, we use the resourced XML files to
build a default database and then set the converted flag. This should solve
the first case.