Closes bug ID 3588982
For some recipes (and I am not quite sure what the pattern is),
mashStep->endTemp_c() is being set to 0C or 32F. Not sure why. I made a small
adjustment so that if the endTemp_c() is 0, use the stepTemp_c() instead. It
isn't perfect, but it seems to work better.