No bug ID in specific.
Chasing down some oddities with the tree views. It would appear that the
filter is either paying attention to only the rowsAboutToBeRemoved() signal or
otherwise ignoring the rowsRemoved() signal. This causes problems when
removing items from the trees, because the proxy's indexes are out of whack
with the models. This should fix that by slotting the model's rowsRemoved and
rowsInserted signals to a slot that invalidates the filter.

Also cleans up the TreeModel code to better reflect what's going on.
insertRows has been replaced by insertRow and the loadTreeModel() code has
been cleaned up to reflect this.