Finally fixed the tree model/view stack corruption bug!
Turns out, there was no slot to catch when an item had changed its name, so the list was not resorting when a name changed,
or when a new item was added, because we first create a blank item, then change its name.

Therefore, we had unnecessary signals linked to invalidate sorting filters and other proxy models, which is not how they
are intended to be used. This caused stack corruption and segfaulting.

Just giving the tree model the ability to pay attention for item name changes and removing the unnecessary signal/slot
connections resolved the whole shebang.

I'm happy :)