The second step for making mashs work.
There was a bug previously that caused every recipe to create at least two
mashs on import: an empty mash when newRecipe() was called and another mash
when the mash was read from teh XML. That has been fixed.

This checkin also changes some of the logic on loading mashs. If the mash has
a name and that name does not already exist in the mash table, the display is
set true. Otherwise, the mash doesn't show.