Closes bug ID 3563845.
This was a much larger issue than it may have at first appeared.

The first problem was that we were creating numerous mashes when importing a
recipe -- newRecipe() created a mash, then mashFromXml() created another mash
and somebody else created a mash. That behaviour has been fixed. This also
necessitated creating a new default database, which I think I messed up and
will need to do another check in.

The second problem was that we were not properly marking mashes as not
displayed. The default for a mash is now to not be displayed, and only when
you save the mash does the display toggle. 

The third problem was that we weren't properly searching the database to see
if we had loaded that mash already. Only the first instance of the name (case
sensitive, which causes a bit of a surprise) is now marked as visible.

The fourth problem was when we copied the mash, we never copied the mashsteps.
This was fixed by defining a new method Database::duplicateMashSteps() since
this was needed in two places.

The fifth problem was that we never loaded the not-displayed mashes from the
database, which caused the named mash to not show anything. After a lot of
thinking about it, I modified the search parameters such that everything is
loaded from the database, except what has been marked as deleted. It is now up
to the views to decide what they want to show and required some serious
reworking of all the SortFilterProxies.

Finally (whew), I had to replace the old mashComboBox with the modern button +
combo. This required reworking some signals, defining a few new classes,
monkeying around with the main window layout, etc.  It was also the easiest
part of it.