Closes bug ID 3522899.
We know pay attention to hop utilization rates. This required exposing the
field in the equipment editor, changing the IBU calculations to respect the
utilization, modifying the schema to default the field to 100.0 instead of 0.0
and updating the default database.

I'm also including a small bit of work so you can mark a piece of equipment as
"default" which will cause that piece of equipment to be selected for any new
recipe.