Signals, signals everywhere. The root cause was that when an equipment profile was selected, we also connected the changed signal. When the save button was pressed, the first field to be updated generated a changed() signal. This signal caused the editor to reload all of the fields using the values from the record in the database. On new equipment, the values in the database are 0. This overwrote everything but the one field that generated the signal.
My fix is to (a) disconnect that signal. We are in a modal window editing the equipment. The only thing that can change
the equipment is the editor we have open. I have reconnected all of the edit boxen signals such that they properly
reformat the input. Finally, the checkboxen should all be functional again.

Mea maxima culpa for this one.