Closes bug 1121200
This was harder than expected. In essence, a change in how we calculated the
total theoretical sugars caused us to change what was stored in the database
from what we converted. This lead to problems with imported data, since the
number made no sense.

This patch fixes three things, and adds a fourth. First, we now store the same
data as we did in previous versions. Second, we makes sure no recalculations
are done when data is migrated. Third, the order was slightly rearranged to
make sure the projected volumes are set before we calculate the projected
points.

The fourth point is that we have to clean up the mess I made. To this end,
MainWindow checks the values for projected points and the efficiencies. If the
points are < 15 or the efficiencies are < 10, we recalculate them. This seems
to clean up my databases.