I hope this makes the windows install and upgrades work. It is done in two parts. The first part is to customize the NSIS.template.in file and point our installer there. I've added some functionality to prompt the user, after installation, for the location of the old databases if a specific registry key isn't found. If the user provides it, we check for a sqlite database. If it exists, that gets copied. If it doesn't, we start going after the XML files. When it's all done, it pokes the location of the database (APPDATA\brewtarget) into the registry. It also pokes a flag into the registry for the next step
When brewtarget starts, it checks to see if the flag is defined. If it is, it
performs the last step of copying the info from options.xml into the registry
and does its thing. Since we copied the files before brewtarget started, the
standard upgrade processing *should* work for both sqlite files and XML files.