I think the database is updated for the new column. I think that needs testing too.
I can create arbitarily deep folder structures based on string input. It
needs cleaning up, but the structure is there. I should be able to search for
a folder too. The search is not quite functional, since it won't work on a
folder like /a/a/a/a/a.

Next step is to make sure I can load from database, and ot make sure the
search works right. That needs to be done with something like /a/b/c/recipe
and /a/b/c/recipe01 and /a/b/recipe02.