I can read a folder attribute from the database and build an arbitrarily deep tree. I can also build mixed folders, containing subtrees and recipes. I can do this all by reading the folder attribute from the database.
The migration stuff has been updated, at rocketman's hints. It still doesn't
quite work. brewtarget crashes with a corrupted stack. If the user selects
"recover" on restart, it seems to work fine. This obviously needs fixing.

The remaining work is now to get the folders exposed to the UI. That's going
to be ... not fun.