If you right click on a recipe that's in a folder and say "New recipe", the new recipe is created in the same folder. If you right click on a folder and say "New recipe", it gets added to that folder.
I still need to test the other create methods.

Create and delete now work for recipes. Now I just need to worry about "move".
And that should be easy, as long as I can trap the right pieces of
information.

Once this is done, I need to consider how to abstract this work out so I do
not end up cut'n'pasting a lot of code.