Still can't figure out what delete means. I've pushed most of the work a layer deeper. The prompts ask if you want to delete the folder and not each element in the folder, but we still delete each element in the folder.
Added a nice bit to expand any folder that the recipe is in when
setTreeSelection is called in MainWindow.

I think I will add a method like firstElement() that will find the first
non-folder element in a tree and return the index. This will (maybe?) make
deletes a bit more please to my eye.