Unit system refactor
I completely rebuilt the Unit and Scales mechanisms.

All of the UnitSystem classes have been pretty much reduced to their
constructor, a new method called loadMap(), thicknessUnit() and unitType().
All of the really hard work is now down in UnitSystem.cpp and unit.cpp.

Unit now exposes getUnit as a static public method. The logic is really hard
to follow, so I've tried to add comments. Basically, if you enter "5 gal", it
will look for the Unit System that matches your current system-level default.
All of the Unit classes have a new method called boundary(), which is used
later.

UnitSystem now has the only implementation of qstringToSI and displayAmount.
By using the QMap populated by the sundrey loadMaps(), displayAmount is
very elegant. If a specific scale is requested, it uses that. Otherwise, it
loops through the map and finds the appropriate scale.