Fixes no unit and surprise entry
Pretty simple change to fix both. The annoying part is that I had to make sure
each UnitSystem knows its default Unit. Not bad, but annoying. After that, it
was a simple change in brewtarget.cpp to make sure the proper unit was
provided to qstringToSI.

If you enter an amount with no unit, it will be interpreted as the default
unit. For example, assume you have SI volumes as the system default and US gallons
selected on the "Calculated batch size" field. If you enter "5" into that
field, it will be interpreted as 5 US gallons.

If you have selected Imperial gallons on a field and have US gallons as your
default, entering 5 gal into the field will be interpreted as 5 imperial
gallons. I think this violates POLA just a little less.