MR120: calories per 330mL for SI units
Fixes a minor annoyance with calories/12oz

Basically, it is always calories per 12oz, even when the volume system is SI.
Given that a "standard" can in europe is 33cL, and at the request of a
european user, I have modified the calories per 12oz to be volume unit
sensitive.

I should really consider doing this as a BtVolumeLabel. The issue is that
there is no real toSI and fromSI for this field. I would need to either
special case it in the BtLineEdit code or do something else.