MR121: fix crash when double-clicking recipes on brewdate
Bug #1416916 Crash on column select

Working the bug. I think this is a hold over from the days when trees had just
one column. When I implemented multi-column trees, I didn't update some logic.
I am still testing.

This closes 1416916

I believe this was a boundary check that did not expand when the bounds did.
Back in the day, there was only ever one column in the trees, so getting any
column > 0 was an error. Once I added more columns, this should have been
updated.

I now check against the number of columns for the tree. I have also added some
protection in the filter so that anything that returns an invalid index will
not be displayed.