MR123: Print output uses units and scales now
Same as another branch, but from a different starting point. All of the print
formats now use the selected units&scales. Whenever possible, I pulled which
unit&scale from the main recipe panel, except when doing something like
printing brewnotes.

To get this to work with an existing recipe and the brewday, you will need to
regenerate the instructions.