Refactored and extracted logging facilities.
I had two problems with the current state of Brewtarget. First, the log
file was pretty useless when trying to debug error because  it's used
in too few places and does not store major events as to used when
looking at the user flow.

The second was that I am in thre process of trying to remove all OS
specific processing from Brewtarget.h and Brewtarget.cpp. I think this
class shouldn't care on which platform it runs and should also be way
clearer as it is one of the central classes.

Here is what was done.
  - Extracted all logging functionalities whilst keeping the same wrapper
    in Brewtarget.
  - Created a Log class to handle all logging.
  - Log is now using RAII on all stream and file descriptors.
  - Log is also now thread-safe.
  - Added INFO and DEBUG case for future inclusion in the application.
  - Log now stores a timestamp
  - Log now use a centralize template.
  - Log also calls a signal on all log entry.
  - The MainWindow now connects to this signal to update the statusBar.

Fixed the header I copy-pasted from another class.

Integration of comments from Phillip. Friend class + private const.

  - Now Log has a private construtor.
  - Brewtarget is a friend class of Log to activate the constructor.
  - Destructor code simplified by a tip from Phillip.