Change translation README to reflect new translation workflow
Now we are keeping the .po files from launchpad in version control,
as launchpad does not provide a versioning mechanism. We are also
generating 2 commits for each translation update. One commit is
only for the imported .po files, and the other is after running
lupdate. This will give us a complete history of the .po files in
case we need to revert anything on launchpad.