Introduction of the user-dir cli option.
This commit introduces the possibility of running brewtarget on a custom
database folder which will not impact your normal usage of the
Application.

This is intended for people with multiple version of Brewtarget and for
people helping test and debug the application.

  * Modification to the settings made while debugging the application no
    longuer interfere with normal settings.
  * Introduction of the --user-dir cli option which allows the caller to
    point to a directory containg the DB. This option will not store
    itself in the settings. Meaning running brewtarget again without param
    will revert to the known user data directory.
  * There were a lot of "file and directory arithmetics" implicated in
    this modification. Things like adding slash, double slash, backward
    slash, ... at the end of paths, and between path element. I stored the
    user data dir as a QDir instead of a QString to eliminate all these
    potential errors. It meant changing calls in a few file.

Fixed Linux version. Was forgotten because of define.

Added explanation on the choice of different app.name