Small script updates
Realized I needed to reset the indices after the import, so added code to do
that.

Started working on my own database, and realized there were problems with
fermentDate field. Fixed that, but I don't like the fix. Instead of using the
date when displayed, it is using the date when the query was run. This could
have some weird side effects.