No cores and initialization works
In reverse order of the summary. I was finally able to get the
QSqlDatabase::tables method to return properly. I made a beginner's mistakes
and confused pass-by-copy with pass-by-reference. Both databases now
initialize properly on startup, and don't initialize when the tables exist.

The cores were a more interesting problem. It would seem PostgreSQL doesn't
like to be closed and/or removed. The driver itself properly takes care of
that when it is destroyed. Or so it seems. I modified unload() so it only
works for SQLite, and the cores went away.