Options dialog made
Added a new tab to the options dialog for configuring the rdbms. If you change
the settings, you cannot save until you successfully test the connection. That
required me writing some code to ... well, test the connection.

This work flow is annoyingly chatty. I pop way too many dialogs, but telling
the user what went wrong is important, as is feed back on it going right.
Hopefully, somebody else will weigh in on this.

We need to think about merging the directory tab with the database tab. They
sort of configure the same thing, so I think they are the same idea. I think I
can do that -- enable only those fields that are relevant to that rdbms, with
a few text changes on the labels should get it done.