Postgres 9.5 and changes
Right now, there is an annoying error message when creating a brewnote.
postgres 9.4 and earlier to not support any kind of "insert or update" syntax
-- frequently called "upsert" in postgres terminology. Postgresql 9.5 does
have an upsert syntax but postgresql 9.5 was released Jan 6, 2016. It isn't
wide spread yet. Unless somebody really screams, I am saying we support 9.5
and forward.

This meant I had to install a new server, which lead me to wondering how I was
going to copy the data, which lead me to modifying convertSqliteToPostgres()
to simply be convertToPostgres(), which lead me to refactoring that code a few
times over. So instead of solving the original bug, I fixed something
completely different and will fix the first bug tomorrow.

I also updated the instructions to remove the perl references.