Big changes
Lots of things got jammed into this.

It mostly works now. You can create recipes and brewnotes now, and it all
looks good. I need to test creating other objects, but that should all work
now.

I changed the database conversion from a post-process to a pre-process. It
made more sense, but it took a lot of effort. The cast majority of this change
is doing that. I really want to try simply re-opening the db to see if we can.

We don't save the database password by default any more. If you decide to save
it, a warning is popped saying you are doing so at your own risk. If you
choose not to, we should prompt at start up for the password. I haven't tested
that.

DatabaseSchemaHelper has been updated to provide useful feed back if something
fails. Every create method. That hurt.

I need to get the updates working next and the merges.